WIP Don't use symlinks in node_modules
Some node packages like babel traverse the complete node_modules tree to
find dependencies i.e. for plugins. However the absolute and resolved
path to the module is used thus no node_modules folder can be found in
the parent folder and the dependencies can't be found.

Due to this, lots of dependencies of downstream dependencies are simply
wrong. In our case, some babel packages won't find their plugins but
without using symlinks they to.

This hack ensures symlinks are only used for dependencies to keep the
overall space consumption low. In the package we're interested in
building, all symlinks will be dereferenced and the contents copied to
the main node_modules directory.

TODO: generation of template package nix code (top)